requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
alabaster==0.7.13
    # via sphinx
astropy==5.2.2
    # via -r requirements.in
babel==2.17.0
    # via sphinx
batman-package==2.5.2
    # via -r requirements.in
beautifulsoup4==4.13.4
    # via furo
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
contourpy==1.1.1
    # via matplotlib
corner==2.2.1
    # via -r requirements.in
cycler==0.12.1
    # via matplotlib
dill==0.4.0
    # via -r requirements.in
docutils==0.20.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
emcee==3.1.6
    # via -r requirements.in
fonttools==4.57.0
    # via matplotlib
furo==2024.8.6
    # via -r requirements.in
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.5.0
    # via sphinx
importlib-resources==6.4.5
    # via matplotlib
jinja2==3.1.6
    # via sphinx
kiwisolver==1.4.7
    # via matplotlib
markupsafe==2.1.5
    # via jinja2
matplotlib==3.7.5
    # via
    #   -r requirements.in
    #   corner
numpy==1.24.4
    # via
    #   -r requirements.in
    #   astropy
    #   batman-package
    #   contourpy
    #   emcee
    #   matplotlib
    #   pandas
    #   pyerfa
packaging==25.0
    # via
    #   astropy
    #   matplotlib
    #   sphinx
pandas==2.0.3
    # via -r requirements.in
pillow==10.4.0
    # via matplotlib
pyerfa==2.0.0.3
    # via astropy
pygments==2.19.1
    # via
    #   furo
    #   sphinx
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via
    #   babel
    #   pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   astropy
requests==2.32.3
    # via sphinx
six==1.17.0
    # via python-dateutil
snowballstemmer==3.0.1
    # via sphinx
soupsieve==2.7
    # via beautifulsoup4
sphinx==7.1.2
    # via
    #   -r requirements.in
    #   furo
    #   sphinx-basic-ng
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-basic-ng==1.0.0b2
    # via furo
sphinx-rtd-theme==3.0.2
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
typing-extensions==4.13.2
    # via beautifulsoup4
tzdata==2025.2
    # via pandas
urllib3==2.2.3
    # via requests
zipp==3.20.2
    # via
    #   importlib-metadata
    #   importlib-resources
